+++ src/shared/ui/tokens.ts
@@ -1,0 +1,2 @@
+export { spacing, radius, zIndex, shadows, colors, tokens, lightTokens } from '@/constants/tokens';
+export type { Tokens } from '@/constants/tokens';
